requirements.txt
dnspython==2.2.1
email-validator==1.2.1
idna==3.4
install==1.3.5
motor==3.0.0
pydantic==1.10.2
pymongo==4.2.0
typing-extensions==4.3.0
uvicorn==0.18.3
fastapi==0.85.0
python-dotenv==0.21.0
pytest==7.1.3
jwt==1.3.1